admin_panel_service/requirements.txt
aiogram==2.12.1
configparser
PyMySQL
cryptography
pycryptodome
TgCrypto
